string class stuff, temporarily adding another base*-file in common
git-svn-id: http://eathena.googlecode.com/svn/branches/eapp@4746 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/common/socket.cpp
@@ -1720,7 +1720,7 @@ int null_console_parse(char *buf)
 int start_console(void) {
 	
 	SOCKET sock=0; // default zero socket, is not used regulary
-	size_t fd;
+	int fd;
     
 	// insert the socket to the fields and get the position
 	fd = SessionInsertSocket(sock);
